Added precursor of snaegel, void_sphere and GravLens enemy object
Prototype of the skill "Snaegel" added and controlled from MenuManager by accessing the speed variable but resulting in undesired behaviour of object skipping or jumping ahead because of  lerping calculation of fraction of journey covered which takes into account the speed variable. No solutions found yet as of now.

Enemy object void_sphere and GravLens along with their script that rotates them and detects player collision is also added.

P.S. The process of making them in Blender is gruelling af.